# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: VAN API
  version: 1.0.0
  description: >
    API REST de VAN para la app móvil (Flutter). Toda la lógica de negocio
    vive en el servidor; la app es un cliente delgado.

    Autenticación: Bearer JWT. El login devuelve access_token (corto,
    30 min) y refresh_token (opaco, 30 días, rotativo). Cuando el
    access_token expira, la app debe llamar POST /auth/refresh con el
    refresh_token para obtener un par nuevo. El refresh_token se revoca
    con POST /auth/logout.

    Envelope de respuestas:
      Exito:  {"success": true,  "data": {...}}
      Error:  {"success": false, "error": {"code": "CODE", "message": "..."}}
...
paths:
...
  /drivers/location:
    post:
      tags: [driver]
      summary: Conductor envía su ubicación actual
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat: {type: number, minimum: -90, maximum: 90}
                lng: {type: number, minimum: -180, maximum: 180}
      responses:
        '200':
          description: Ubicación actualizada
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error', description: NOT_ONLINE }
  /drivers/online:
    post:
      tags: [driver]
      summary: Toggle online/offline del conductor
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                is_online: {type: boolean, default: false}
      responses:
        '200':
          description: Estado actualizado
  /drivers/nearby:
    get:
      tags: [driver]
      summary: Conductores online + libres cerca de una posición
      security:
        - bearerAuth: []
      parameters:
        - {name: lat, in: query, required: true, schema: {type: number}}
        - {name: lng, in: query, required: true, schema: {type: number}}
        - {name: radius, in: query, schema: {type: number, default: 10}, description: km}
        - {name: vehicle_type, in: query, schema: {type: string, enum: [moto, auto]}}
      responses:
        '200':
          description: Lista de conductores cercanos ordenados por distancia
        '400': { $ref: '#/components/responses/Error' }
  # ─── A4: Driver config ───
  /driver/accepted-payments:
    get:
      tags: [driver]
      summary: Métodos de pago aceptados por el conductor
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de métodos aceptados
    put:
      tags: [driver]
      summary: Actualizar métodos de pago aceptados
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accepted_payments:
                  type: array
                  items: {type: string, enum: [efectivo, mercadopago, transferencia, tarjeta, billetera]}
      responses:
        '200':
          description: Métodos actualizados
  /driver/qr:
    post:
      tags: [driver]
      summary: Subir imagen de QR de MercadoPago (base64)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image]
              properties:
                image: {type: string, description: Base64 o data URL}
      responses:
        '200':
          description: QR subido
        '400': { $ref: '#/components/responses/Error' }
  /driver/payment-methods:
    get:
      tags: [driver]
      summary: Listar métodos de cobro del conductor
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de métodos de cobro
    post:
      tags: [driver]
      summary: Agregar método de cobro
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: {type: string, enum: [card, mercadopago, transfer]}
                details: {type: object}
      responses:
        '201':
          description: Método creado
  /driver/payment-methods/{method_id}:
    delete:
      tags: [driver]
      summary: Eliminar método de cobro
      security:
        - bearerAuth: []
      parameters:
        - name: method_id
          in: path
          required: true
          schema: {type: integer}
      responses:
        '200':
          description: Eliminado
        '404': { $ref: '#/components/responses/Error' }
  # ─── A5: Geocode + Reviews + Favorites + Voucher ───
  /geo/geocode:
    get:
      tags: [geo]
      summary: Geocodificación Nominatim (proxy)
      security:
        - bearerAuth: []
      parameters:
        - {name: q, in: query, required: true, schema: {type: string}, description: Dirección a buscar}
      responses:
        '200':
          description: Coordenadas
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          lat: {type: number}
                          lng: {type: number}
                          display_name: {type: string}
        '404': { $ref: '#/components/responses/Error', description: No encontrado }
  /users/{user_id}/reviews:
    get:
      tags: [auth]
      summary: Reseñas de un usuario
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema: {type: integer}
        - {name: role, in: query, schema: {type: string, enum: [driver, passenger]}, description: Filtrar por rol del rater}
      responses:
        '200':
          description: Lista de reseñas
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        success: {type: boolean, example: false}
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_CREDENTIALS
            message:
              type: string
              example: Credenciales inválidas
...
    OkResponse:
      type: object
      properties:
        success: {type: boolean}
        data:
          type: object
          properties:
            ok: {type: boolean}
...